refactor getInputDatasetSchema into src
