+++ bin/check-input-datasets.js
@@ -1,48 +1,14 @@
 const fsPromises = require("fs").promises;
-const Ajv = require("ajv").default;
-
+const getInputDatasetSchema = require("../src/data-validation/getInputDatasetSchema");
 const { DATA_FOLDER_NAME } = require("../src/process-single-dataset/constants");
-const {
-  readDatasetJson
-} = require("../src/utils");
+const { readDatasetJson } = require("../src/utils");
 const dataPrep = require("../src/data-validation/data-prep");
-const unpackInputDataset = require("../src/process-single-dataset/steps/unpack-input-dataset")
+const unpackInputDataset = require("../src/process-single-dataset/steps/unpack-input-dataset");
 
 // referenced partial schemas
-const refSchemas = [
-  require("../src/data-validation/schema/definitions.schema.json"),
-  require("../src/data-validation/schema/discrete-feature-options.schema.json"),
-  require("../src/data-validation/schema/array-items/discrete-feature-option.schema.json"),
-  require("../src/data-validation/schema/array-items/file-info.schema.json"),
-  require("../src/data-validation/schema/array-items/input-measured-features.schema.json"),
-];
-
-// Schemas that describe the handoff files that define a dataset or a
-// collection of datasets.
-const inputMegaset = require("../src/data-validation/schema/input-megaset.schema.json");
-const inputDatasetInfo = require("../src/data-validation/schema/input-dataset-info.schema.json");
-const inputMeasuredFeatures = require("../src/data-validation/schema/input-measured-features-doc.schema.json");
-const inputDataSet = require("../src/data-validation/schema/input-dataset.schema.json");
-const inputImages = require("../src/data-validation/schema/images.schema.json");
-const featureDef = require("../src/data-validation/schema/feature-def.schema.json");
-
-const INPUT_DATASET_SCHEMA = [
-  ...refSchemas,
-  inputImages,
-  featureDef,
-  inputMegaset,
-  inputDatasetInfo,
-  inputMeasuredFeatures,
-  inputDataSet,
-];
 const INPUT_DATASET_SCHEMA_FILE = "input-dataset.schema.json";
 const INPUT_MEGASET_SCHEMA_FILE = "input-megaset.schema.json";
-
-const ajv = new Ajv({
-  coerceTypes: true,
-  removeAdditional: true,
-  schemas: INPUT_DATASET_SCHEMA,
-});
+const ajv = getInputDatasetSchema;
 
 const checkForError = (fileName, json, schemaFileName) => {
   const { valid, error } = dataPrep.validate(
@@ -70,7 +36,6 @@ const checkForError = (fileName, json, schemaFileName) => {
   }
 };
 
-
 const validateDatasets = () => {
   fsPromises
     .readdir(`./${DATA_FOLDER_NAME}`)
@@ -91,11 +56,15 @@ const validateDatasets = () => {
 
       for (const name of files) {
         try {
-          const subFiles = await fsPromises.readdir(`./${DATA_FOLDER_NAME}/${name}`);
+          const subFiles = await fsPromises.readdir(
+            `./${DATA_FOLDER_NAME}/${name}`
+          );
           if (subFiles.includes("dataset.json")) {
             foldersToCheck.push(`./${DATA_FOLDER_NAME}/${name}`);
           }
-        } catch (error) {console.log(error)}
+        } catch (error) {
+          console.log(error);
+        }
       }
       for (const datasetFolder of foldersToCheck) {
         const topLevelJson = await readDatasetJson(datasetFolder);
@@ -125,7 +94,8 @@ const validateDatasets = () => {
         console.log("\x1b[31m");
         throw Error("Validation failed");
       }
-    }).catch(console.log)
+    })
+    .catch(console.log);
 };
 
 validateDatasets();
