+++ src/data-validation/full-schema.js
@@ -50,4 +50,5 @@ module.exports = {
     iDataset: ajv.getSchema("input-dataset.schema.json"),
     iDatasetInfo: ajv.getSchema("input-dataset-info.schema.json"),
     iMegaset: ajv.getSchema("input-megaset.schema.json"),
+    
 }
